feat(model): support latest model version trigger
