+++ config/model/task_input.json
@@ -12,7 +12,6 @@
           "image/*"
         ],
         "title": "Image base64",
-        "pattern": "^[A-Za-z0-9+/]*={0,2}$",
         "type": "string"
       }
     },
@@ -34,7 +33,6 @@
           "image/*"
         ],
         "title": "Image base64",
-        "pattern": "^[A-Za-z0-9+/]*={0,2}$",
         "type": "string"
       }
     },
@@ -56,7 +54,6 @@
           "image/*"
         ],
         "title": "Image base64",
-        "pattern": "^[A-Za-z0-9+/]*={0,2}$",
         "type": "string"
       }
     },
@@ -78,7 +75,6 @@
           "image/*"
         ],
         "title": "Image base64",
-        "pattern": "^[A-Za-z0-9+/]*={0,2}$",
         "type": "string"
       }
     },
@@ -100,7 +96,6 @@
           "image/*"
         ],
         "title": "Image base64",
-        "pattern": "^[A-Za-z0-9+/]*={0,2}$",
         "type": "string"
       }
     },
@@ -122,7 +117,6 @@
           "image/*"
         ],
         "title": "Image base64",
-        "pattern": "^[A-Za-z0-9+/]*={0,2}$",
         "type": "string"
       }
     },
@@ -243,7 +237,6 @@
         ],
         "instillUIOrder": 2,
         "title": "Image base64",
-        "pattern": "^[A-Za-z0-9+/]*={0,2}$",
         "type": "string"
       },
       "prompt": {
